requirements.txt
evadb==0.3.9
numpy
pandas
transformers
google-generativeai
torch
scikit-learn
sentence-transformers